# apps/admin/src/api/generated/users/model/task.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * New Year Game API
 * OpenAPI spec version: 1.0.0
 */

export interface Task {
  answer: string
  id: string
  meaning: string
  question: string
  startDate: string
  wordIndex: number
}