# apps/admin/src/api/generated/users/model/updateTextRequest.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * New Year Game API
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateTextRequest {
  /** @nullable */
  endDate?: string | null
  /** @nullable */
  startDate?: string | null
  text?: string
  title?: string
}