# apps/admin/src/api/generated/users/model/userBasic.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * New Year Game API
 * OpenAPI spec version: 1.0.0
 */

export interface UserBasic {
  correctAttempts: number
  id: number
  middleName: string
  name: string
  phone: string
  surname: string
}