# apps/admin/src/api/generated/users/model/userDetailed.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * New Year Game API
 * OpenAPI spec version: 1.0.0
 */
import type { UserDetailedAnswersItem } from './userDetailedAnswersItem'
import type { UserDetailedReviews } from './userDetailedReviews'

export interface UserDetailed {
  answers?: UserDetailedAnswersItem[]
  division?: string
  email?: string
  id?: number
  lotteryNumber?: number
  middleName?: string
  name?: string
  phone?: string
  place?: string
  reviews?: UserDetailedReviews
  surname?: string
}